shop/requirements.txt
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.4.17
chardet==3.0.4
cntk==2.0
configparser==3.5.0
decorator==4.1.2
Django==1.10.7
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.8.1
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.4
django-redis==4.8.0
django-tables2==1.10.0
django-treebeard==4.1.2
django-widget-tweaks==1.4.1
enum34==1.1.6
factory-boy==2.8.1
Faker==0.7.18
flake8==3.3.0
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.5
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.15
mccabe==0.6.1
mock==2.0.0
olefile==0.44
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
phonenumbers==8.6.0
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
psycopg2==2.7.3
ptyprocess==0.5.2
purl==1.3.1
pycodestyle==2.3.1
pycountry==17.1.8
pyflakes==1.5.0
Pygments==2.2.0
PyMySQL==0.7.11
pyprof2calltree==1.4.1
pysolr==3.6.0
python-dateutil==2.6.1
pytz==2017.2
rcssmin==1.0.6
redis==2.10.5
requests==2.18.1
rjsmin==1.0.12
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.4a1
sqlparse==0.2.3
traitlets==4.3.2
transifex-client==0.12.4
Unidecode==0.4.21
urllib3==1.21.1
uWSGI==2.0.15
wcwidth==0.1.7
Werkzeug==0.12.2
whitenoise==3.3.0
Whoosh==2.7.4
